+++ win.h
@@ -53,7 +53,7 @@ extern "C" {
 /* Flags for portable clock_gettime call. */
 #define CLOCK_REALTIME  0
 
-void*   mmap(void *addr, size_t len, int prot, int flags, int fildes, off_t off);
+void*   mmap(void *addr, size_t len, int prot, int flags, int fildes, ssize_t off);
 int     munmap(void *addr, size_t len);
 int     mprotect(void *addr, size_t len, int prot);
 int     msync(void *addr, size_t len, int flags);
